refactor(groceries): extract derivation helpers
